Refactor orchestrator.py into orchestrators/ package
Split the monolithic orchestrator module into separate files for each
orchestrator class, improving navigability and maintainability.
