+++ scripts/run_call.py
@@ -53,7 +53,7 @@ from rumil.calls.call_registry import (
 from rumil.calls.prioritization import run_prioritization
 from rumil.database import DB
 from rumil.models import CallStage, CallType, FindConsiderationsMode
-from rumil.orchestrator import create_root_question
+from rumil.orchestrators import create_root_question
 from rumil.settings import Settings, get_settings, _settings_var
