+++ tests/test_investigate.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from rumil.orchestrator import Orchestrator
+from rumil.orchestrators import Orchestrator
 
 
 @pytest.mark.integration
@@ -13,8 +13,11 @@ async def test_investigate_creates_pages(tmp_db, question_page):
     await orch.run(question_page.id)
 
     rows = (
-        await tmp_db.client.table("pages").select("id").eq("run_id", tmp_db.run_id).execute()
+        await tmp_db.client.table("pages")
+        .select("id")
+        .eq("run_id", tmp_db.run_id)
+        .execute()
     )
     page_ids = [r["id"] for r in rows.data]
     new_ids = [pid for pid in page_ids if pid != question_page.id]
-    assert len(new_ids) >= 1, f"Expected new pages, only found the original question"
+    assert len(new_ids) >= 1, "Expected new pages, only found the original question"
